backend/requirements.txt
# Core Framework
Flask==3.1.0
Flask-CORS==5.0.0
Flask-JWT-Extended==4.7.1
python-dotenv==1.0.1

# Database
SQLAlchemy==2.0.36
Flask-SQLAlchemy==3.1.1
PyMySQL==1.1.1
cryptography==42.0.5
alembic==1.14.0

# API & Validation
marshmallow==3.23.2
flask-restx==1.3.0

# ML & Data Science
pandas==2.2.3
numpy==2.2.1
scikit-learn==1.6.1
prophet==1.1.6
nltk==3.9.1

# Utilities
python-dateutil==2.9.0
joblib==1.4.2

# Testing
pytest==8.3.4
pytest-cov==6.0.0

# Security
bcrypt==4.2.1

# AI/LLM
openai==1.58.1

# File Processing
pdfplumber==0.11.4
openpyxl==3.1.5
python-docx==1.1.2
chardet==5.2.0

# SMS Notifications
twilio==9.0.4

# OCR (Optional - requires Tesseract installed on system)
# pytesseract==0.3.13
# pillow==10.4.0
